use -std=c++11 or -std=gnu++11

+++ VecArrayStat.h
@@ -5,7 +5,7 @@
 
 class ArrayStat {
 private:
-	std::multiset<int> myset;
+	std::multiset<double> myset;
 public:
     ArrayStat(const char *file_name);
